Code cleanup for PR.
Remove possible existing dprintf

We defined dprintf in spl, so a simpler version can be used there,
but when ZFS headers are loaded, the proper dprintf is used.

Simply remove any existing dprintf before changing to avoid
warnings.

Fixed changed to Linux that were unintended

Move Windows specific things from common headers into os/

+++ lib/libspl/include/os/windows/dirent.h
@@ -254,7 +254,7 @@ __internal_opendir(wchar_t *wname, int size)
 	if (!data)
 		goto out_of_memory;
 	wname[extra_prefix + size - 1] = 0;
-	data->fd = (int)CreateFileW(wname, 0, 0, NULL, OPEN_EXISTING,
+	data->fd = (int)(uintptr_t)CreateFileW(wname, 0, 0, NULL, OPEN_EXISTING,
 	    FILE_FLAG_BACKUP_SEMANTICS, 0);
 	wname[extra_prefix + size - 1] = L'\\';
 	data->count = 16;
